add some filters

+++ app/src/main/java/com/juandou/mediapikcer/util/VideoScanHelper.kt
@@ -14,7 +14,7 @@ import java.io.File
 import java.io.FileOutputStream
 
 object VideoScanHelper {
-    private val MEDIA_PICK_TEMP_FOLDER_PATH = Environment.getExternalStorageDirectory().absolutePath + "/mediaPicker/temp";
+    private val MEDIA_PICK_TEMP_FOLDER_PATH = Environment.getExternalStorageDirectory().absolutePath + "/mediaPicker/temp/";
     fun start(context: Context, handler: Handler) {
         Thread {
             doScan(context, handler)
@@ -37,7 +37,6 @@ object VideoScanHelper {
 
         try {
             cursor = context.contentResolver.query(MediaStore.Video.Media.EXTERNAL_CONTENT_URI, mediaColumns, null, null, sortOrder)
-
             cursor?.moveToFirst().apply {
                 when (this) {
                     true -> parseData(context, cursor, handler)
@@ -80,9 +79,8 @@ object VideoScanHelper {
                     }
                 }
 
-
-                thumbNailPath = thumbNailPath.isNullOrEmpty().let {
-                    generateThumbNail(filePath)
+                if (thumbNailPath.isNullOrEmpty()) {
+                    thumbNailPath = generateThumbNail(filePath)
                 }
 
                 videoList.add(MediaData(id, createTime, duration, albumName, filePath, thumbNailPath, mimeType, null, null))
@@ -100,27 +98,30 @@ object VideoScanHelper {
     }
 
     private fun generateThumbNail(filePath: String?): String? {
-        filePath.isNullOrEmpty().let {
+        filePath?.let {
             var out: FileOutputStream? = null
             try {
-                val folderArr: List<String> = filePath!!.split(File.separator)
-                val fileName = folderArr.isNotEmpty().let {
-                    folderArr[folderArr.size - 1]
+                val folderArr: List<String>? = filePath.split(File.separator)
+                val fileName = folderArr?.let {
+                    folderArr[folderArr.size - 1] + ".jpg"
                 }
+
                 val folder = File(MEDIA_PICK_TEMP_FOLDER_PATH)
-                folder.exists().not().let {
+                if (!folder.exists()) {
                     folder.mkdirs()
                 }
-                val file = File(MEDIA_PICK_TEMP_FOLDER_PATH + fileName)
-                file.exists().not().let {
-                    val bitmap = createVideoThumbnail(filePath)
-                    out = FileOutputStream(file)
-                    bitmap.compress(Bitmap.CompressFormat.JPEG, 50, out)
-                    out!!.flush()
-                    bitmap.recycle()
-                }
-                return file.absolutePath
 
+                val file = File(MEDIA_PICK_TEMP_FOLDER_PATH.plus(fileName))
+                if (!file.exists()) {
+                    val bitmap = createVideoThumbnail(filePath)
+                    bitmap?.let {
+                        out = FileOutputStream(file)
+                        bitmap.compress(Bitmap.CompressFormat.JPEG, 50, out)
+                        out?.flush()
+                        bitmap.recycle()
+                    }
+                    return file.absolutePath
+                }
             } catch (e: Exception) {
             } finally {
                 out?.close()
@@ -129,7 +130,7 @@ object VideoScanHelper {
         return null
     }
 
-    private fun createVideoThumbnail(filePath: String?): Bitmap {
+    private fun createVideoThumbnail(filePath: String): Bitmap? {
         var bitmap: Bitmap? = null
         val retriever = MediaMetadataRetriever()
         try {
@@ -156,6 +157,6 @@ object VideoScanHelper {
                 bitmap = Bitmap.createScaledBitmap(bitmap, w, h, true)
             }
         }
-        return bitmap as Bitmap
+        return bitmap
     }
 }
